feat: use custom errors
Starting from Solidity [v0.8.4](https://blog.soliditylang.org/2021/04/21/custom-errors/), there is a convenient and gas-efficient way to explain to users why an operation failed through the use of custom errors. This way is less expensive, especially when it comes to deploy cost, than passing a string to require or assert.