Add Komorebi No Sekai Project
Based on ERC721A from Azuki team, we implemented custom and unique features in our smart contract. We use Chainlink VRF to randomly select a winner on-chain that will receive a set of predefined NFTs. Holders of our KNS NFTs will be able to participate actively to the creation of the manga. Hence the community will have the opportunity to collaborate directly with the artist and the team. In addition, each holder will be assigned to a side: Furaribi or Naito Raito. The side can be used later in the roadmap.